[REF] po-lint: Disable 'puncspacing'
In English you say *the 'Configuration' section.* and in Spanish *la sección ''Configuración'.* So, in english, after the last `'` character we have a space but in spanish we don't.

In almost all the other warnings, we have grammar and syntax conflcits like that, but, in this module at least, `puncspacing` didn't help me fix another "real" translation problem. So, I give it the useless award this time.