[FIX] manifest-version-format: Don't ignore valid Odoo versions
Currently, the parameter --valid_odoo_versions is being ignored when
checking a module's version in its manifest. So, for example, if the
only valid version is 11.0, and a module has in its manifest
`10.0.1.0.0`, this lint wouldn't be triggered, resulting in a false
negative.

This commit ensures the above parameter is taken into account when
checking module versions set in the manifest.

A new test case is provided for this behavior