[FIX] missing-manifest-dependency: Exclude tests directory
Since tests files are executed only when running tests, external
dependencies defined on those files should not be required in the
manifest as external Python dependencies. Those dependencies will not be
used in a non-testing environment and, if tests are run and dependencies
are not satisfied, tests will fail anyway.

This commit causes all files into the tests directory to be ignored by
the lint.

This is similar to: https://github.com/OCA/pylint-odoo/pull/75