Update ES style conventions
- Allow bitwise operators.
- Allow implicit coercion when using `~`. See https://github.com/getify/You-Dont-Know-JS/blob/master/types%20&%20grammar/ch4.md#the-curious-case-of-the- to know why it's useful.
- Require inline block spacing.
- Require multiline dangling commas. Nobody cares about IE8.
- Limit line width to 80 characters, except for the `odoo.define()` one, which tends to require extra space.
- Require spaces after commas and comments.
- Require consistent indentation of 4 spaces.
- Require consistent line breaks inside braces.
- Require consistent newlines in object property declarations.
- Requre `"use strict"`.
- Require sorting imports.
- Require consistent spacing around blocks and keywords.
- Comments must start with upper case.
- Block comments require a space before.
- Relax some rules for jsdoc comments, but warn if it is missing for any declared method. See below for details.

About the jsdoc relaxation, the need comes from the fact that sometimes you're extending another object, and thus you don't need to reexplain everything if it is explained in the parent class.

A common coding pattern is:

```js
return this._super.apply(this, arguments);
```

In such cases, it should be enough to declare the return type, since further descriptions should be found in the upstream class.

Another common pattern is reusing upstream parameters, or using parameters with names that are specific enough so as to not require a description, so the requirement of param description is relaxed here too.

In any case, remember that when you're overriding a method, you can just do a jsdoc like this:

```js
/**
 * @override
 */
overriden_method: function (param1, param2) {
    // Some logic here
    return this._super.apply(this, arguments);
}
```

The `@override` keyword above removes any further requirements of manually copying anything from upstream's jsdoc. Use it!