[IMP] manifest-required-author: Allow multiple possible required authors
There are cases when the required autor for modules may be one among
several ones, e.g. the author is required, but it is not always the
same. For cases like those, it is useful to support multiple values for
this lint.

This commit improves the lint so it accepts multiple values separated by
commas.

Test cases are provided to cover one and multiple values passed to the
parameter to set required authors.