Allow console statements
These are required in Odoo in order to mark a tour as failed manually.

It can also be helpful to log some interesting details for developers that shouldn't matter for users. Odoo uses it, in fact.

See https://github.com/OCA/website/pull/627#discussion_r285006941 where this was discovered, and https://github.com/OCA/pylint-odoo/pull/244#issuecomment-493358163 where the need is further explained.