[REF]  odoo-addons-relative-import: does not apply in tests/ directory (#307)
Fix https://github.com/OCA/pylint-odoo/issues/297

It is more correct to use absolute imports in tests.
Tests are not truly constitutive of the module, they just happen to be stored inside the module.

Usually, in Python libs, they're stored outside of the package. Importantly, using the absolute import proves that the addon we develop works as a library too.