[REF] Improved repo structure to facilitate testing The existence of src was not necessary, and made it difficult to run unittests directly, since the imports depended on pylint_odoo being installed in the system, instead of importing directly from source code, so tox was the only way to run tests (AFAIK).
After this change, unittests can be directly run without
tox, this makes the development process faster and
integrates very well with IDEs that let you run and
debug unit tests directly, plus it removes something
that is not needed.

pylint itself uses this structure, so I think that
is another reason in favor of this change.