[FIX] Exclude missing-return checks for computed methods
Fixes #450. Compute methods are not supposed to return any value,
even if there is a super() in it, since super() shouldn't return
anything either.

The only reliable way to ensure a field is computed is if has
been declared in the same python module (file), so this exclusion
only applies to a limited set of cases.