[ADD] deprecated-odoo-model-method: New check to detect deprecated method for each Odoo version (#470)
* [ADD] deprecated-odoo-model-method: New check to detect deprecated method for each Odoo version (#470)

This new check looks for model methods that have been deprecated and
should not be overriden in model classes. It takes into account odoo
versions and comes with sane defaults which can be replaced by a custom
list in the configuration file.

Closes #469.

* [REF] tests: support custom rcfiles

Tests can now specify their own rcfiles, previously it was always
/dev/null (no rcfile). This makes it possible to test the behavior of
tests when options are provided through rcfiles.

There is also no need to supply /dev/null as a rcfile. Pylint
internally uses a default RC file if not provided. It also makes it
redundant to use `_add_rcfile_default_pylintrc` in tests.