[REF]: update to pylint 3.0.0, drop support for py3.7 (#474)
Pylint 3.0.0 has been released and with it a bunch of fixes and
improvements. This repository has been updated to depend on this new
version.

BREAKING CHANGE: Pylint 3.0.0 drops support for Python 3.7 so this
plugin no longer supports Python 3.7 as well.

* update code to work with pylint v3

Pylint v3.0.0 changed some of its behaviour externally (to users) and
internally as well (for developers) by depending on a newer release of
astroid. Code has been updated to work with said changes and fix the
current test suite.

Specifically astroid has dropped support for visit_tryexcept and other
friends, replacing it by visit_try.

Pylint's invalid check no longer checks for a  minimum length of three
characters as well, so the expected messages had to be updated.

Closes #473.